attribute editor and element selector
