+++ packages/gui/src/components/primitives/index.ts
@@ -4,3 +4,4 @@ export * from './ColorPicker'
 export * from './UnitSelect'
 export * from './ValueSelect'
 export * from './ThemeValue'
+export * from './Combobox'
